% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSRAreads.R
\name{assignSRAreads}
\alias{assignSRAreads}
\title{Assign SRA reads for desposited 10X single cell RNAseq datasets}
\usage{
assignSRAreads(
  working_dir = NULL,
  input_dir = NULL,
  outdir = NULL,
  parallel = FALSE
)
}
\arguments{
\item{working_dir}{Set working directory. Provide absolute paths}

\item{input_dir}{Set to directory containing fasterq-dump or fastq-dump files}

\item{outdir}{Directory to output the read legnths as SRRXXXXX_X.readlength.txt}

\item{parallel}{if \emph{TRUE}, uses mclapply from parallel package. \emph{default=FALSE}}
}
\value{
A dataframe containing SRR_ID, assigned_read, orig_names, new_names and cellranger_names
A csv will be written into the outdir "assigned_SRAreads.csv".
}
\description{
This function takes an input directory containing fasterq-dump or fastq-dump files.
Fasterq and fastq outputs must be generated from using the split option.
Example:
fasterq-dump --include-technical --split-files
or
fastq-dump --split-files
\strong{PLEASE NOTE} This will not take the output of fasterq/fastq-dump if you specified the --type argument
}
\details{
250 reads from all fastq files are then sampled from the listed fastqs and assigned to R1, R2 or I3.
10,000 reads will also be sampled if the reads are matched for R1
}
