% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSRAreads.R
\name{renameAll}
\alias{renameAll}
\title{Rename all fastas to the correct filenames}
\usage{
renameAll(assigned_SRA = NULL, input_dir = NULL, format = NULL)
}
\arguments{
\item{assigned_SRA}{Input the output of \emph{assignSRAreads}}

\item{input_dir}{Input the same input_dir argument as used for \emph{assignSRAreads}}
}
\value{
The "assigned_SRAreads.csv" file will be updated with a name_check column containing
the boolean outcome of listed fastq names matched to the intended new_names column.
}
\description{
This function takes the output of \emph{assignSRAreads} and directly changes the filename in place.
}
